requirements.txt
opencv_python==4.4.0.42
tensorflow_gpu==2.3.0
Flask==1.1.4
Keras==2.4.3
Flask_Cors==3.0.8
requests==2.22.0
numpy==1.18.5
matplotlib==3.0.0
Pillow
tensorflow==2.3.0